// models/BuyAssetRequestModel.ts
/**
 * dfs-api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 2022-04-19T11:06:16Z
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class BuyAssetRequestModel {
    'symbol': string;
    'amountToSpend'?: number;
    'tokenToBuy'?: number;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "symbol",
            "baseName": "symbol",
            "type": "string",
            "format": ""
        },
        {
            "name": "amountToSpend",
            "baseName": "amountToSpend",
            "type": "number",
            "format": ""
        },
        {
            "name": "tokenToBuy",
            "baseName": "tokenToBuy",
            "type": "number",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return BuyAssetRequestModel.attributeTypeMap;
    }

    public constructor() {
    }
}