// models/GetAvaliableAssetsResponseModelBody.ts
/**
 * dfs-api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 2022-04-19T11:06:16Z
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { GetAvaliableAssetsResponseModelBodyItems } from './GetAvaliableAssetsResponseModelBodyItems';
import { HttpFile } from '../http/http';

export class GetAvaliableAssetsResponseModelBody {
    'items'?: Array<GetAvaliableAssetsResponseModelBodyItems>;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "items",
            "baseName": "items",
            "type": "Array<GetAvaliableAssetsResponseModelBodyItems>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return GetAvaliableAssetsResponseModelBody.attributeTypeMap;
    }

    public constructor() {
    }
}