// models/GetAvaliableAssetsResponseModelBodyItems.ts
/**
 * dfs-api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 2022-04-19T11:06:16Z
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class GetAvaliableAssetsResponseModelBodyItems {
    'tokenAmmount': number;
    'price24HoursAgo': number;
    'currentPrice': number;
    'id': string;
    'time': Date;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "tokenAmmount",
            "baseName": "tokenAmmount",
            "type": "number",
            "format": "float"
        },
        {
            "name": "price24HoursAgo",
            "baseName": "price24HoursAgo",
            "type": "number",
            "format": "float"
        },
        {
            "name": "currentPrice",
            "baseName": "currentPrice",
            "type": "number",
            "format": "float"
        },
        {
            "name": "id",
            "baseName": "_id",
            "type": "string",
            "format": ""
        },
        {
            "name": "time",
            "baseName": "time",
            "type": "Date",
            "format": "date-time"
        }    ];

    static getAttributeTypeMap() {
        return GetAvaliableAssetsResponseModelBodyItems.attributeTypeMap;
    }

    public constructor() {
    }
}