// models/GetAvaliableBalanceResponseModel.ts
/**
 * dfs-api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 2022-04-19T11:06:16Z
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { GetAvaliableBalanceResponseModelBody } from './GetAvaliableBalanceResponseModelBody';
import { HttpFile } from '../http/http';

export class GetAvaliableBalanceResponseModel {
    'message': string;
    'body'?: GetAvaliableBalanceResponseModelBody;
    'statusCode'?: number;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "message",
            "baseName": "message",
            "type": "string",
            "format": ""
        },
        {
            "name": "body",
            "baseName": "body",
            "type": "GetAvaliableBalanceResponseModelBody",
            "format": ""
        },
        {
            "name": "statusCode",
            "baseName": "statusCode",
            "type": "number",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return GetAvaliableBalanceResponseModel.attributeTypeMap;
    }

    public constructor() {
    }
}