// models/GetAvaliableBalanceResponseModelBody.ts
/**
 * dfs-api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 2022-04-19T11:06:16Z
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { GetAvaliableBalanceResponseModelBodyItem } from './GetAvaliableBalanceResponseModelBodyItem';
import { HttpFile } from '../http/http';

export class GetAvaliableBalanceResponseModelBody {
    'item'?: GetAvaliableBalanceResponseModelBodyItem;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "item",
            "baseName": "item",
            "type": "GetAvaliableBalanceResponseModelBodyItem",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return GetAvaliableBalanceResponseModelBody.attributeTypeMap;
    }

    public constructor() {
    }
}