// models/GetAvaliableBalanceResponseModelBodyItem.ts
/**
 * dfs-api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 2022-04-19T11:06:16Z
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class GetAvaliableBalanceResponseModelBodyItem {
    'reference': string;
    'uid': string;
    'newBalance': number;
    'oldBalance': number;
    'difference': number;
    'time': Date;
    'type': string;
    'userId': string;
    'transactionId': string;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "reference",
            "baseName": "reference",
            "type": "string",
            "format": ""
        },
        {
            "name": "uid",
            "baseName": "uid",
            "type": "string",
            "format": ""
        },
        {
            "name": "newBalance",
            "baseName": "newBalance",
            "type": "number",
            "format": "float"
        },
        {
            "name": "oldBalance",
            "baseName": "oldBalance",
            "type": "number",
            "format": "float"
        },
        {
            "name": "difference",
            "baseName": "difference",
            "type": "number",
            "format": "float"
        },
        {
            "name": "time",
            "baseName": "time",
            "type": "Date",
            "format": "date-time"
        },
        {
            "name": "type",
            "baseName": "type",
            "type": "string",
            "format": ""
        },
        {
            "name": "userId",
            "baseName": "userId",
            "type": "string",
            "format": ""
        },
        {
            "name": "transactionId",
            "baseName": "transactionId",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return GetAvaliableBalanceResponseModelBodyItem.attributeTypeMap;
    }

    public constructor() {
    }
}